requirements.txt
homeassistant==2026.1.1
beautifulsoup4==4.12.3
lxml==5.3.0
josepy
pycryptodome>=3.19.0
qrcode>=7.4.0
requests>=2.31.0